+++ crates/eww/src/widgets/widget_definitions.rs
@@ -345,10 +345,12 @@ const WIDGET_NAME_CHECKBOX: &str = "checkbox";
 fn build_gtk_checkbox(bargs: &mut BuilderArgs) -> Result<gtk::CheckButton> {
     let gtk_widget = gtk::CheckButton::new();
     def_widget!(bargs, _g, gtk_widget, {
+        // @prop checked - whether the checkbox is toggled or not when created
         // @prop timeout - timeout of the command. Default: "200ms"
         // @prop onchecked - action (command) to be executed when checked by the user
         // @prop onunchecked - similar to onchecked but when the widget is unchecked
-        prop(timeout: as_duration = Duration::from_millis(200), onchecked: as_string = "", onunchecked: as_string = "") {
+        prop(checked: as_bool = false, timeout: as_duration = Duration::from_millis(200), onchecked: as_string = "", onunchecked: as_string = "") {
+            gtk_widget.set_active(checked);
             connect_signal_handler!(gtk_widget, gtk_widget.connect_toggled(move |gtk_widget| {
                 run_command(timeout, if gtk_widget.is_active() { &onchecked } else { &onunchecked }, &[] as &[&str]);
             }));
@@ -513,6 +515,18 @@ fn build_gtk_button(bargs: &mut BuilderArgs) -> Result<gtk::Button> {
     Ok(gtk_widget)
 }
 
+/// @var icon-size - "menu", "small-toolbar", "toolbar", "large-toolbar", "button", "dnd", "dialog"
+fn parse_icon_size(o: &str) -> Result<gtk::IconSize> {
+    enum_parse! { "icon-size", o,
+        "menu" => gtk::IconSize::Menu,
+        "small-toolbar" | "toolbar" => gtk::IconSize::SmallToolbar,
+        "large-toolbar" => gtk::IconSize::LargeToolbar,
+        "button" => gtk::IconSize::Button,
+        "dnd" => gtk::IconSize::Dnd,
+        "dialog" => gtk::IconSize::Dialog,
+    }
+}
+
 const WIDGET_NAME_IMAGE: &str = "image";
 /// @widget image
 /// @desc A widget displaying an image
@@ -530,7 +544,12 @@ fn build_gtk_image(bargs: &mut BuilderArgs) -> Result<gtk::Image> {
                 let pixbuf = gtk::gdk_pixbuf::Pixbuf::from_file_at_size(std::path::PathBuf::from(path), image_width, image_height)?;
                 gtk_widget.set_from_pixbuf(Some(&pixbuf));
             }
-        }
+        },
+        // @prop icon - name of a theme icon
+        // @prop icon-size - size of the theme icon
+        prop(icon: as_string, icon_size: as_string = "button") {
+            gtk_widget.set_from_icon_name(Some(&icon), parse_icon_size(&icon_size)?);
+        },
     });
     Ok(gtk_widget)
 }
@@ -875,29 +894,44 @@ fn build_gtk_label(bargs: &mut BuilderArgs) -> Result<gtk::Label> {
 
     def_widget!(bargs, _g, gtk_widget, {
         // @prop text - the text to display
+        // @prop truncate - whether to truncate text (or pango markup). If `show-truncated` is `false`, or if `limit-width` has a value, this property has no effect and truncation is enabled.
         // @prop limit-width - maximum count of characters to display
         // @prop truncate-left - whether to truncate on the left side
-        // @prop show-truncated - show whether the text was truncated
+        // @prop show-truncated - show whether the text was truncated. Disabling it will also disable dynamic truncation (the labels won't be truncated more than `limit-width`, even if there is not enough space for them), and will completly disable truncation on pango markup.
         // @prop unindent - whether to remove leading spaces
-        prop(text: as_string, limit_width: as_i32 = i32::MAX, truncate_left: as_bool = false, show_truncated: as_bool = true, unindent: as_bool = true) {
-            let limit_width = limit_width as usize;
-            let char_count = text.chars().count();
-            let text = if char_count > limit_width {
-                let mut truncated: String = if truncate_left {
-                    text.chars().skip(char_count - limit_width).collect()
+        prop(text: as_string, truncate: as_bool = false, limit_width: as_i32 = i32::MAX, truncate_left: as_bool = false, show_truncated: as_bool = true, unindent: as_bool = true) {
+            let text = if show_truncated {
+                // gtk does weird thing if we set max_width_chars to i32::MAX
+                if limit_width == i32::MAX {
+                    gtk_widget.set_max_width_chars(-1);
                 } else {
-                    text.chars().take(limit_width).collect()
-                };
-                if show_truncated {
-                    if truncate_left {
-                        truncated.insert_str(0, "...");
-                    } else {
-                        truncated.push_str("...");
-                    }
+                    gtk_widget.set_max_width_chars(limit_width);
                 }
-                truncated
-            } else {
+                if truncate || limit_width != i32::MAX {
+                    if truncate_left {
+                        gtk_widget.set_ellipsize(pango::EllipsizeMode::Start);
+                    } else {
+                        gtk_widget.set_ellipsize(pango::EllipsizeMode::End);
+                    }
+                } else {
+                    gtk_widget.set_ellipsize(pango::EllipsizeMode::None);
+                }
+
                 text
+            } else {
+                gtk_widget.set_ellipsize(pango::EllipsizeMode::None);
+
+                let limit_width = limit_width as usize;
+                let char_count = text.chars().count();
+                if char_count > limit_width {
+                    if truncate_left {
+                        text.chars().skip(char_count - limit_width).collect()
+                    } else {
+                        text.chars().take(limit_width).collect()
+                    }
+                } else {
+                    text
+                }
             };
 
             let text = unescape::unescape(&text).context(format!("Failed to unescape label text {}", &text))?;
@@ -905,7 +939,30 @@ fn build_gtk_label(bargs: &mut BuilderArgs) -> Result<gtk::Label> {
             gtk_widget.set_text(&text);
         },
         // @prop markup - Pango markup to display
-        prop(markup: as_string) { gtk_widget.set_markup(&markup); },
+        // @prop truncate - whether to truncate text (or pango markup). If `show-truncated` is `false`, or if `limit-width` has a value, this property has no effect and truncation is enabled.
+        // @prop limit-width - maximum count of characters to display
+        // @prop truncate-left - whether to truncate on the left side
+        // @prop show-truncated - show whether the text was truncatedd. Disabling it will also disable dynamic truncation (the labels won't be truncated more than `limit-width`, even if there is not enough space for them), and will completly disable truncation on pango markup.
+        prop(markup: as_string, truncate: as_bool = false, limit_width: as_i32 = i32::MAX, truncate_left: as_bool = false, show_truncated: as_bool = true) {
+            if (truncate || limit_width != i32::MAX) && show_truncated {
+                // gtk does weird thing if we set max_width_chars to i32::MAX
+                if limit_width == i32::MAX {
+                    gtk_widget.set_max_width_chars(-1);
+                } else {
+                    gtk_widget.set_max_width_chars(limit_width);
+                }
+
+                if truncate_left {
+                    gtk_widget.set_ellipsize(pango::EllipsizeMode::Start);
+                } else {
+                    gtk_widget.set_ellipsize(pango::EllipsizeMode::End);
+                }
+            } else {
+                gtk_widget.set_ellipsize(pango::EllipsizeMode::None);
+            }
+
+            gtk_widget.set_markup(&markup);
+        },
         // @prop wrap - Wrap the text. This mainly makes sense if you set the width of this widget.
         prop(wrap: as_bool) { gtk_widget.set_line_wrap(wrap) },
         // @prop angle - the angle of rotation for the label (between 0 - 360)